third_party/WebKit/LayoutTests/platform/win/editing/pasteboard/drop-text-events-expected.txt

This tests that Drag drop fires textInput events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS actualTextEventData is 'PlainHello'
PASS testTargetTextarea.value is 'PlainHello'
PASS actualTextEventData is 'PlainHello'
PASS testTargetInput.value is 'PlainHello'
FAIL actualTextEventData should be  (of type string). Was null (of type object).
FAIL 0 <= testTargetEditable.innerHTML.indexOf('PlainHello') should be true. Was false.
PASS actualTextEventData is 'RichHello'
PASS testTargetTextarea.value is 'RichHello'
PASS actualTextEventData is 'RichHello'
PASS testTargetInput.value is 'RichHello'
PASS actualTextEventData is ''
PASS 0 <= testTargetEditable.innerHTML.indexOf('<b>Rich</b>Hello') is true
FAIL textInputCount should be 6. Was 5.
PASS actualTextEventData is 'PlainHello'
PASS testTargetTextarea.value is ''
PASS actualTextEventData is 'PlainHello'
PASS testTargetInput.value is ''
FAIL actualTextEventData should be  (of type string). Was null (of type object).
PASS testTargetEditable.innerHTML is 'placeholder'
PASS actualTextEventData is 'RichHello'
PASS testTargetTextarea.value is ''
PASS actualTextEventData is 'RichHello'
PASS testTargetInput.value is ''
PASS actualTextEventData is ''
PASS testTargetEditable.innerHTML is 'placeholder'
FAIL textInputCount should be 6. Was 5.
PASS successfullyParsed is true

TEST COMPLETE
